fixed ldflags in Makefile, fixed version command
